pyproject.toml
[project]
name = "lang-ollama"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "googlesearch-python>=1.3.0",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.21",
    "langchain-milvus>=0.1.9",
    "langchain-ollama>=0.3.2",
    "praw>=7.8.1",
    "pydantic>=2.11.3",
    "pymilvus>=2.5.6",
    "sentence-transformers>=4.1.0",
    "sumy>=0.11.0",
    "tqdm>=4.67.1",
    "underthesea>=6.8.4",
]
